Manual Mode with Timers Working 20150312 SLM NEW TimerSetEvent & TimerSetListener - Event to be fired when the duration changed, so list item duration could be set.
NEW BUT NOT IMPLEMENTED SCountdownTimer - wanted to create a timer that
counts down rather than up. This is not yet implemented.

NEW STimeTextView - Created to try to separate the display functionality
("00:00") from the runnable in STimerView. But still have some
co-mingling of functionality because I needed a way to fire an event
when the value of the duration changed so that the list item would be
updated 
	
RENAME SLMTimerView to STimerView. Also made this be derived from
STimeTextView, trying to separate the runnable functionality from the
simple display (i.e. "00:00").

CustomListAdapter - Many changes to organize the Adapter & ViewHolder
classes. Added cirDuraionSTTV and cirDurationMs, removed listItem.
position holds the position in the list so really didn't need to drag
around the extra baggage of the ListItem in the ViewHolder. Modified
most of the Tags to be set to the holder variable rather than the
position. This allowed me to get rid of the newHolder variable. Just
working to be able to change the duration variable in the list items so
that we could maintain the correct value for the Duration that is
displayed / set on the screen. This is not quite working but I'm close.

ListItem - Added displayDuration and logDuration and removed duration.

TimerOickerDialogFragment - Mods to try to update the duration when the
timer picker OK button was pressed. This is not quite working but I'm
close.


20150217 SLM
NEW - TimerPickerDialogFragment - new Picker widget to allow user to
enter minutes and seconds to run a sprinkler circuit.

ListItem - added a SetTimer button such that when pressed, the
TimerPickerDialogFragment will pop up so user can set Timer time.

list_item - added set_timer button to xml file

CustomListAdapter - mods to allow SetTimer button to be pressed and
responded to (Temp response for now). 

20150216 (and before) SLM	
ListItem - Changed displayTime to duration. startTime and stopTime will
be used for logging purposes and will include date and time info.
duration will only hold the value for the number of ms that the circuit
has run.

NEW - SLMTimerView - new widget to replace Chronometer since Chronometer
only counts up and I want to be able to count down as well. Extends
TextView.

CustomListAdapter - Mods to deal with new SLMTimerView instead of
Chronometer. holder is a pain to work with:-)

NEW - CircuitOnOffEvent - added to work with EventBus

NEW - TimePickerFragment - added to handle setting duration of manual
mode operation.Not implemented yet.

NEW - OspiGetResultsListener and OspiResultsListener - added / separated
out to handle JSoup requests.
 OnResults() not implemented yet.

Added use of greenrobot EventBus library.