SLM Comments for checkin 04/16/2015    Basic changes to get a crude version of the Schedule creating activity working.    20150415 SLM ScheduleActivity - Added Lifecycle methods which added an EventBus register & unregister calls. Added command type to OspiGetResultsAsyncTask calls.
ManualModeActivity - Added command type to OspiGetResultsAsyncTask
calls.
		
OspiGetResultsAsyncTask - Added a second argument to constructor to
indicate the type of command (a GET which expects a result, and an
EXECUTE, which sends to request but does not expect a result (for my cp
or change program requests))
- Added if case to handle different command types

20150331 SLM TimePickerDialogFragment - add code to allow for specifying
24-hour time format.

NEW ChangeProgramEvent - use to create a cp or change program request to
the Ospi web app (currently just creates a new program with pid
hardcoded to -1.

20150323-26 SLM

	NEW TimePickerDialogFragment - added to handle starttime setting on
ScheduleActivity screen.
	
NEW ScheduleActivity - new Activity to display the setup for a schedule.
This is just a temporary attempt to get "something" working.

NEW ManualModeActivity  - MainActivity RENAMED to ManualModeActivity and
make minor changes so it can be called up by a new MainActivity that
displays 2 buttons: one for Manual Mode and one for the Schedule.

MainActivity - MODIFIY to only have 2 buttons: one to display the
ManualModeActivity and one to display the ScheduleActivity.

NEW schedule_setup.xml - define screen elements for ScheduleActivity.

main.xml - RENAMED to manual_mode.xml

main.xml MODIFIED for new MainActivity

list_item.xml - minor mods to spacing and/or dimens changes

AndroidManifest.xml - added new Activities. Currently using explicit
intents for manual mode and schedule.

dimens.xml, strings.xml - mods as needed.
style.xml - mods as needed.